scripts: Address shellcheck issues
Report issues by shell files with:

    git ls-files '*.sh' | xargs shellcheck --exclude=SC2162

Address the following:

- SC1087: Use braces when expanding arrays
- SC1091: Not following
- SC2004: `$`/`${}` is unnecessary on arithmetic variables
- SC2024: `sudo` doesn't affect redirects
- SC2034: foo appears unused. Verify it or export it
- SC2086: Double quote to prevent globbing and word splitting
- SC2087: Quote `EOF`
- SC2115: Use `"${var:?}"` to ensure this never expands to `/*`
- SC2148: Add a shebang

Addresses (at least partially) some POSIX/dash issues:

- SC2113: `function` keyword is non-standard
- SC3010: In POSIX sh, `[[` `]]` is undefined
- SC3014: In POSIX sh, `==` in place of `=` is undefined
- SC3020: In POSIX sh, `&>` is undefined
- SC3046: In POSIX sh, `source` in place of `.` is undefined

Does not address:

- SC2162: `read` without `-r` will mangle backslashes
- Any other POSIX/dash-specific issues

+++ scripts/flash.sh
@@ -14,7 +14,6 @@ then
   echo "model '${MODEL}' not found" >&2
   exit 1
 fi
-MODEL_DIR="$(realpath "models/${MODEL}")"
 
 DMI_MODEL="$(cat /sys/class/dmi/id/product_version)"
 if [ "${DMI_MODEL}" != "${MODEL}" ]
@@ -29,7 +28,7 @@ export BASEDIR="system76-firmware-update"
 # Clean build directory
 mkdir -p build
 BUILD="$(realpath "build/${MODEL}")"
-rm -rf "${BUILD}/${BASEDIR}"
+rm -rf "${BUILD:?}/${BASEDIR}"
 mkdir -p "${BUILD}/${BASEDIR}"
 
 # Rebuild and copy firmware-update
@@ -50,7 +49,7 @@ fi
 
 # Locate EFI partition mount path
 EFI_PATH="$(bootctl --print-esp-path)"
-if [ -z "${EFI_PATH}" -o ! -d "${EFI_PATH}" ]
+if [ -z "${EFI_PATH}" ] || [ ! -d "${EFI_PATH}" ]
 then
     echo "EFI system partition '${EFI_PATH}' not found" >&2
     exit 1
@@ -63,7 +62,7 @@ then
     echo "EFI system partition name not found" >&2
     exit 1
 fi
-EFI_PART="$(cat /sys/class/block/${EFI_PART_NAME}/partition)"
+EFI_PART="$(cat "/sys/class/block/${EFI_PART_NAME}/partition")"
 
 # Locate EFI disk
 EFI_DISK=""
